Auth Feature Implemented

+++ server/migrations/0001_create-initial-tables.sql
@@ -9,6 +9,6 @@ CREATE TABLE IF NOT EXISTS todo (
    description      VARCHAR(1000)                           NOT NULL,
    created_at       TIMESTAMP WITHOUT TIME ZONE default current_timestamp,
    name             VARCHAR(1000)   DEFAULT 'superuser',
-   user_id          NUMERIC         DEFAULT -1              NOT NULL,
+   user_id          INTEGER         DEFAULT -1              NOT NULL,
    CONSTRAINT todo_pk PRIMARY KEY (id)
 );
